+++ example/lib/pages/home_page.dart
@@ -9,7 +9,12 @@ class HomePage extends ECSWidget {
   @override
   Widget build(BuildContext context, ECSContext ecs) {
     ecs.onEnter(() {
+      debugPrint('HomePage entered');
       ecs.get<ReloadUserEvent>().trigger();
+    });
+
+    ecs.onExit(() {
+      debugPrint('HomePage exited');
     });
 
     return Scaffold(